Finally got everythin working.
- Degree of freedom calculation no longer is done
as global variable

- the normal grid can now be calculated with the
call of a function

- the forces grid can now be calculated with the
call of a function

- The Enviroment now uses the 4 dimensional
matrix as design space, as well as as Observation
space

- The Enviroment can now be selected to run on
two different modes, one for training and one for
evaluation. The training mode will continuously
generate random load and boundary conditions to
train the agent. The evaluation mode will use
a fixed set of load and boundary conditions to
evaluate the agent.

- The rendering Method used to display the
progress of the agent has been updated to
display the grid in a more readable way
as well as adding a heatmap for compliance forces

- Added multiple different small helper functions